src/requirements.txt
argparse
pandas 
gremlinpython
azure.cosmos
typing